[dev.cc] runtime: fix linux build
TBR=austin
CC=golang-codereviews
https://golang.org/cl/176760044

+++ src/runtime/os1_linux.go
@@ -48,8 +48,8 @@ func futexsleep(addr *uint32, val uint32, ns int64) {
 	// is not, even timediv is too heavy, and we really need to use just an
 	// ordinary machine instruction.
 	if ptrSize == 8 {
-		ts.set_sec(ns / 1000000000)
-		ts.set_nsec(ns % 1000000000)
+		ts.set_sec(int32(ns / 1000000000))
+		ts.set_nsec(int32(ns % 1000000000))
 	} else {
 		ts.tv_nsec = 0
 		ts.set_sec(timediv(ns, 1000000000, (*int32)(unsafe.Pointer(&ts.tv_nsec))))
